Update notification mechansim for label CI checks CI check_labels won't fail but leaving a comment, fail merge attempt instead.
